r-car m2: set PREFERRED_VERSION_gstreamer1.0* to 1.2.% when h/w acceleration enabled
The R-Car M2 multimedia acceleration provides acceleration support for GStreamer 1.2.3,
so set the preferred version to 1.2.%. Required for meta-ivi which supports both 1.2.3
and 1.4.x but defaults to 1.4.x.

Both M2 Koelsch and M2 Porter boards are supported, so generalise the previous
Kolesch only setting to all M2 boards. Also make the setting dependent on h/w
acceleration being enabled, else use 1.4.x. This solves a build issue when the
s/w rendered 1.2.x was built when h/w acceleration was not enabled.