gdp: Fix missing gst faad plugin build error when s/w mmp is used.
Only add gfx or mmp to image when h/w acceleration enabled.

This fixes an issue reported on list [1] where the GDP do_rootfs fails:
'ERROR: gstreamer1.0-plugins-bad-faad not found in the base feeds
(koelsch cortexa15hf-vfp-neon cortexa15hf-vfp armv7ahf-vfp-neon
armv7ahf-vfp armv6hf-vfp armv5ehf-vfp armv5hf-vfp noarch any all)'

The issue was introduced with meta-renesas commit 632f4b7e7c05db
which integrated a mechanism to add gfx and multimedia packages
as needed to the IMAGE. Faad requires it be added to the image
dependencies and is found in meta-openembedded/meta-multimedia/.
By default the R-Car M2 Koeslch board local.conf and bblayers.conf
in GDP do not enable h/w multimedia acceleration and so those
requirements were not met. Resulting in the error as nothing
was causing a Gst 1.4.x Faad plug-in to be built.

However for GDP and the general case the BSP should only be adding
packages it requires when h/w acceleration is enabled. Therefore
modify the mechanism to do just that. When s/w rendering is
used it should be left to layers outside of the BSP to do what
is required.

[1] http://lists.genivi.org/pipermail/genivi-projects/2016-January/001360.html