multimedia: Fix packagegroup-lcb-oss-codecs* QA issue (PACKAGES)
When multi-media is enabled in the build for the Low Cost Boards (LCB),
such as M2 Porter and E2 Silk, Yocto can report a QA Issue of the
following form:
ERROR: QA Issue: packagegroup-lcb-oss-codecs is listed in PACKAGES multiple times, this leads to packaging errors. [packages-list]
ERROR: QA Issue: packagegroup-lcb-oss-codecs-dbg is listed in PACKAGES multiple times, this leads to packaging errors. [packages-list]
ERROR: QA Issue: packagegroup-lcb-oss-codecs-dev is listed in PACKAGES multiple times, this leads to packaging errors. [packages-list]
ERROR: QA Issue: packagegroup-lcb-oss-codecs-ptest is listed in PACKAGES multiple times, this leads to packaging errors. [packages-list]

Between Yocto BSP v1.8.0 and v1.10.0 the related information for those
boards in packagegroup-rcar-gen2-multimedia.bbappend was integrated
into the .bb. The bbappend file should have been deleted but was not
and the resulting duplication causes the QA Issue. So delete the file.