u-boot: select wayland config by default
The Genivi Compositor components IVI-Shell and IVI-Extension use Wayland/Weston.
Unlike the general Renesas R-Car Gen 2 Yocto BSP we do not need to support
both X11 and Wayland. As a result set the default u-boot configuration for
the boards to be the one optimised for Wayland.

This eliminates the need to set it explicitly in the local.conf for the
Genivi Baselines and GDP.